+++ tests/sql.rs
@@ -50,6 +50,7 @@ fn select_query_result(id: &str) -> mapepire::QueryResult {
                 },
             ],
             job: None,
+            parameters: vec![],
         },
         data: vec![row],
         cont_id: None,
@@ -58,6 +59,8 @@ fn select_query_result(id: &str) -> mapepire::QueryResult {
         error: None,
         sqlcode: None,
         sqlstate: None,
+        parameter_count: None,
+        output_parms: vec![],
     }
 }
 
@@ -74,6 +77,7 @@ fn dml_query_result(id: &str, count: i64) -> mapepire::QueryResult {
             column_count: 0,
             columns: vec![],
             job: None,
+            parameters: vec![],
         },
         data: vec![],
         cont_id: None,
@@ -82,6 +86,8 @@ fn dml_query_result(id: &str, count: i64) -> mapepire::QueryResult {
         error: None,
         sqlcode: None,
         sqlstate: None,
+        parameter_count: None,
+        output_parms: vec![],
     }
 }